+++ src/main/java/blackjack/model/player/Participant.java
@@ -1,6 +1,6 @@
 package blackjack.model.player;
 
-import blackjack.model.State;
+import blackjack.model.state.State;
 import blackjack.model.card.CardDeck;
 
 import java.util.List;
